added logout and login

+++ backend/php/register.php
@@ -11,18 +11,33 @@
     <form method="post">
         <fieldset>
             <legend>Neuen Account erstellen:</legend>
-            Benutzername:<br>
+            <div id="username">
+                Benutzername:</div>
             <input type="text" name="account">
-            <br>
-            Passwort:<br>
+            <div id="pwd">
+                Passwort:</div>
             <input type="password" name="password">
-            <br><br>
-            Passwort wiederholen:<br>
+            <div id="r_pwd">
+                Passwort wiederholen:
+            </div>
             <input type="password" name="passwordRepeat">
             <br><br>
             <input type="submit" value="Submit">
         </fieldset>
     </form>
+    <script>
+        errors = {};
+
+        function showErrors() {
+            if (errors.user) {
+                document.getElementById("username").innerText += err.user;
+            }
+            if (errors.pwd) {
+                document.getElementById("r_pwd").innerText += err.pwd;
+            }
+        }
+    </script>
+
     <?PHP
 
     if (isset($_POST['account']) && isset($_POST['password']) && isset($_POST['passwordRepeat'])) {
@@ -30,18 +45,24 @@
         $passwd = $_POST['password'];
         $repeated_passwd = $_POST['passwordRepeat'];
 
-        if ($passwd == $repeated_passwd) {
-            include "./salter.php";
-            $account = salter($account);
-            $passwd = salter($passwd);
-            $file = '../data/raw_passwd.csv';
-            $new_line = $account . ',' . $passwd . "\n";
-            if (file_put_contents($file, $new_line, FILE_APPEND | LOCK_EX)) {
-                echo "<script>alert('Registered successfully!')</script>";
+        $file = '../data/raw_passwd.csv';
+        include "./users.php";
+        echo "<script> errors = {}'</script>";
+        if (!userExists($account)) {
+            if ($passwd == $repeated_passwd) {
+                $account = salter($account);
+                $passwd = salter($passwd);
+                $new_line = $account . ',' . $passwd . "\n";
+                if (file_put_contents($file, $new_line, FILE_APPEND | LOCK_EX)) {
+                    echo "<script>alert('Registered successfully!')</script>";
+                }
+            } else {
+                echo "<script> errors['pwd'] = 'Passwörter stimmen nicht überein'</script>";
             }
         } else {
-            echo "<script>alert('Passwörter stimmern nicht überein!')</script>";
+            echo "<script> errors['user'] = 'Benutzername ist bereits vergeben'</script>";
         }
+        echo "<script> if(errors){showErrors()}'</script>";
     }
     ?>
